+++ embc-app/ClientApp/src/app/app-routing.module.ts
@@ -14,6 +14,8 @@ import { VolunteerLoginComponent } from './volunteer-login/volunteer-login.compo
 import { VolunteerDashboardComponent } from './volunteer-dashboard/volunteer-dashboard.component';
 import { EvacueeRegistrationComponent } from './evacuee-registration/evacuee-registration.component';
 import { TesterPageComponent } from './tester-page/tester-page.component';
+import { EvacueeRegistrationOneComponent } from './evacuee-registration/evacuee-registration-one/evacuee-registration-one.component';
+import { EvacueeRegistrationConfirmationComponent } from './evacuee-registration';
 
 const routes: Routes = [
   {
@@ -61,7 +63,26 @@ const routes: Routes = [
   },
   {
     path: 'register-evacuee',
-    component: EvacueeRegistrationComponent
+    component: EvacueeRegistrationComponent,
+    children: [
+      {
+        path: '',
+        redirectTo: 'new',
+        pathMatch: 'full'
+      },
+      {
+        path: 'complete',
+        component: EvacueeRegistrationOneComponent
+      },
+      {
+        path: 'new',
+        component: EvacueeRegistrationOneComponent
+      },
+      {
+        path: 'confirmation',
+        component: EvacueeRegistrationConfirmationComponent
+      }
+    ]
   },
   {
     path: 'register-evacuee/:essFileNumber',
