+++ embc-app/ClientApp/src/app/evacuee-registration/evacuee-registration-confirmation/evacuee-registration-confirmation.component.html
@@ -1,0 +1,3 @@
+<p>
+  evacuee-registration-confirmation works!
+</p>
