+++ embc-app/ClientApp/src/app/evacuee-registration/evacuee-registration-confirmation/evacuee-registration-confirmation.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-evacuee-registration-confirmation',
+  templateUrl: './evacuee-registration-confirmation.component.html',
+  styleUrls: ['./evacuee-registration-confirmation.component.scss']
+})
+export class EvacueeRegistrationConfirmationComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
