+++ embc-app/ClientApp/src/app/evacuee-registration/evacuee-registration-one/evacuee-registration-one.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { EvacueeRegistrationOneComponent } from './evacuee-registration-one.component';
+
+describe('EvacueeRegistrationOneComponent', () => {
+  let component: EvacueeRegistrationOneComponent;
+  let fixture: ComponentFixture<EvacueeRegistrationOneComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ EvacueeRegistrationOneComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(EvacueeRegistrationOneComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
