+++ embc-app/ClientApp/src/app/evacuee-registration/index.ts
@@ -1,0 +1,3 @@
+export * from './evacuee-registration.module';
+export * from './evacuee-registration.component';
+export * from './evacuee-registration-confirmation/evacuee-registration-confirmation.component';
